+++ R/datasets.R
@@ -1,14 +1,14 @@
 
-sp_datasets_i <- tibble::tribble(~id, ~name, ~implemented,
-                               "finm", "FIN 2-12 M - Pln\\u011bn\\u00ed rozpo\\u010dtu M\\u0158O", F,
-                               "finu", "FIN 2-04 U - Pln\\u011bn\\u00ed rozpo\\u010dtu KAP, OSS a SF (2010 - 2014)", F,
-                               "finsf", "FIN 2-04 U - Pln\\u011bn\\u00ed rozpo\\u010dtu SF", F,
-                               "misris", "FIN 1-12 OSS - Pln\\u011bn\\u00ed rozpo\\u010dtu KAP a OSS", F,
-                               "rozv", "Rozvaha", F,
-                               "ppt", "P\\u0159ehled pen\\u011b\\u017en\\u00edch tok\\u016f", F,
-                               "pozvk", "P\\u0159ehled o zm\\u011bn\\u00e1ch vlastn\\u00edho kapit\\u00e1lu", F,
-                               "pril", "P\\u0159\\u00edloha", F,
-                               "vykzz", "V\\u00fdkaz zisk\\u016f a ztr\\u00e1t", F) %>%
+sp_datasets_i <- tibble::tribble(~id, ~name, ~implemented, ~dir,
+                               "finm", "FIN 2-12 M - Pln\\u011bn\\u00ed rozpo\\u010dtu M\\u0158O", F, "FinM",
+                               "finu", "FIN 2-04 U - Pln\\u011bn\\u00ed rozpo\\u010dtu KAP, OSS a SF (2010 - 2014)", F, "FinU",
+                               "finsf", "FIN 2-04 U - Pln\\u011bn\\u00ed rozpo\\u010dtu SF", F, "FinSF",
+                               "misris", "FIN 1-12 OSS - Pln\\u011bn\\u00ed rozpo\\u010dtu KAP a OSS", F, "FinOSS",
+                               "rozv", "Rozvaha", F, "Rozvaha",
+                               "ppt", "P\\u0159ehled pen\\u011b\\u017en\\u00edch tok\\u016f", F, "PenezniToky",
+                               "pozvk", "P\\u0159ehled o zm\\u011bn\\u00e1ch vlastn\\u00edho kapit\\u00e1lu", F, "VlastniKapital",
+                               "pril", "P\\u0159\\u00edloha", F, "Priloha",
+                               "vykzz", "V\\u00fdkaz zisk\\u016f a ztr\\u00e1t", F, "ZiskZtraty") %>%
   dplyr::mutate(name = stringi::stri_unescape_unicode(.data$name)) %>%
   dplyr::arrange(.data$id)
 # stringi::stri_escape_unicode("xxx")
@@ -17,12 +17,12 @@ sp_datasets_i <- tibble::tribble(~id, ~name, ~implemented,
 #'
 #' Contains IDs and names of all available datasets that can be retrieved by get_dataset.
 #'
-#' See <https://monitor.statnipokladna.cz/2019/zdrojova-data/transakcni-data> for a more detailed descriptions
+#' See <https://monitor.statnipokladna.cz/datovy-katalog/transakcni-data> for a more detailed descriptions
 #' of the datasets.
 #'
 #' @format A data frame with 9 rows and 3 variables:
 #' \describe{
-#'   \item{\code{id}}{character. Dataset ID, used as `dataset_id` argument to `get_dataset`.}
+#'   \item{\code{id}}{character. Dataset ID, used as `dataset_id` argument to `sp_get_dataset`.}
 #'   \item{\code{name}}{character. Dataset name, mostly corresponds to title on the statnipokladna GUI.}
 #' }
 #' @family Lists of available entities
@@ -33,13 +33,14 @@ sp_datasets <- sp_datasets_i %>% dplyr::select(.data$id, .data$name) %>%
 get_dataset_url <- function(dataset_id, year = 2018, month = 12, check_if_exists = TRUE) {
   if(!(dataset_id %in% sp_datasets_i$id)) usethis::ui_stop("Not a valid dataset ID")
   dataset_name <- sp_datasets_i[sp_datasets_i$id == dataset_id, "name"]
+  dataset_dir <- sp_datasets_i[sp_datasets_i$id == dataset_id, "dir"]
   usethis::ui_info("Building URL for dataset {usethis::ui_value(dataset_id)}: {usethis::ui_value(dataset_name)}, {usethis::ui_value(stringr::str_c(year,'-',month))}")
-  x <- stringr::str_glue("{sp_base_url}/data/{year}_{month}_Data_CSUIS_{toupper(dataset_id)}.zip")
+  x <- stringr::str_glue("{sp_base_url}/data/extrakty/csv/{dataset_dir}/{year}_{month}_Data_CSUIS_{toupper(dataset_id)}.zip")
   # print(x)
   if(!curl::has_internet()) usethis::ui_stop(c("No internet connection. Cannot continue. Retry when connected.",
                                                "If you need offline access to the data across R sessions, set the {ui_field('dest_dir')} parameter."))
   if(check_if_exists) {
-    iserror <- httr::http_error(x, httr::config(followlocation = 0L), USE.NAMES = FALSE)
+    iserror <- httr::http_error(x)
     if(iserror) usethis::ui_stop("File does not exist for this dataset and period combination.")
   }
   doc_url <- stringr::str_glue("{sp_base_url}/data/struktura/{dataset_id}.xlsx")
@@ -101,9 +102,15 @@ get_dataset_doc <- function(dataset_id, dest_dir = ".", download = TRUE) {
 #' Files are stored in a temp folder as determined by `tempdir()` and further sorted into
 #' subdirectories by dataset, year and month. They persist per session to avoid redownloads.
 #'
+#' How data for different time periods is exported differs by dataset.
+#' This has significant implications for how you get to usable full-year numbers or time series in different tables.
+#' See `vignette("statnipokladna")` for details on this.
+#'
 #' @param dataset_id A dataset ID. See `id` column in `sp_datasets` for a list of available codelists.
 #' @param year year, numeric, 2015-2018 for some datasets, 2010-2019 for others. Defaults to 2018.
+#' (see Details for how to work with data across time periods.)
 #' @param month month, numeric. Must be between 1 and 12. Defaults to 12.
+#' (see Details for how to work with data across time periods.)
 #' @param dest_dir character. Directory in which downloaded files will be stored. Defaults to `tempdir()`. Will be created if it does not exist.
 #' @param redownload Redownload even if file has already been downloaded? Defaults to FALSE.
 #'